MAINTAINER: new maintainer file handling
Define new schema and move to scripts/maintainer
